requirements.txt
authlib>=1.6.6,<2.0.0                                 # For RestAPI Tool
click>=8.1.8,<9.0.0                                   # For CLI tools
fastapi>=0.124.1,<1.0.0                               # FastAPI framework
google-api-python-client>=2.157.0,<3.0.0               # Google API client discovery
google-auth[pyopenssl]>=2.47.0                         # Google Auth library
google-adk                                             # For google.adk.agents.llm_agent import
google-genai                                           # For google.genai module used in agent.py
a2a-sdk>=0.3.4,<0.4.0
litellm>=1.82